refactor(git): move git operations into vcs/git provider
