+++ internal/vcs/git/git.go
@@ -1,7 +1,6 @@
 package git
 
 import (
-	"github.com/mabd-dev/reposcan/internal/gitx"
 	"github.com/mabd-dev/reposcan/internal/vcs"
 	"github.com/mabd-dev/reposcan/pkg/report"
 )
@@ -17,7 +16,7 @@ func (p *Provider) Type() vcs.Type {
 }
 
 func (p *Provider) CheckRepoState(path string) (report.RepoState, []string) {
-	state, warnings := gitx.CheckRepoState(path)
+	state, warnings := CheckRepoState(path)
 	state.VCSType = string(vcs.TypeGit)
 
 	return state, warnings
